+++ tp3/src/tests.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-matrix crear_matriz(int filas, int columnas, float valores[]) {
+matrix crear_matriz(int filas, int columnas, double valores[]) {
     matrix nueva(filas, columnas);
     int c = 0;
     for (size_t i = 0; i < filas; i++) {
@@ -21,7 +21,7 @@ void mostrar(matrix &matriz, u_int fila, u_int cantidadColumnas) {
     std::cout << fila << " tuvieja\n";
     for (unsigned int i = 0; i < cantidadColumnas; i++) {
         for (unsigned int j = 0; j < cantidadColumnas; j++) {
-            float elem = matriz.dame_elem_matrix(fila, i + j * cantidadColumnas);
+            double elem = matriz.dame_elem_matrix(fila, i + j * cantidadColumnas);
             std::cout << ((elem == 0) ? " " : "*") << ' ';
         }
         std::cout << '\n';
@@ -34,7 +34,7 @@ int main(int argc, char **argv) {
 //matrix imagen(argv[1]);
 //imagen.mostrar();
 
-float arr1[] = {1,2,3,4,5,6,7,
+double arr1[] = {1,2,3,4,5,6,7,
                 1,2,3,4,5,6,7,
                 1,2,3,4,5,6,7,
                 1,2,3,4,5,6,7,
@@ -53,7 +53,7 @@ GeneradorRectas::dame_rectas_sobre_base(rectas, 1, 1, alto, ancho);
 matrix destino(rectas.size(), imagen.dame_filas() * imagen.dame_columnas());
 cout << destino.dame_filas() << ", " << destino.dame_columnas();
 
-vector<float> velocidades(rectas.size());
+vector<double> velocidades(rectas.size());
 
 AplicadorRectas::aplicar_rectas(imagen, rectas, velocidades, destino);
 for (int i = 0; i < rectas.size(); ++i){
@@ -65,7 +65,7 @@ matrix velocidades_dis(destino.dame_columnas(),1);
 destino.Cuadrados_Minimos(tiemposMatriz,velocidades_dis);   
 matrix velocidades_ori(49,1);
 velocidades_ori.pasar_matriz_vector(imagen);
-float error = velocidades_ori.ECM(velocidades_dis);
+double error = velocidades_ori.ECM(velocidades_dis);
 cout << "error cuadratico medio es " << error << endl;
 destino.mostrar();
 velocidades_dis.mostrar();
